Refactor - Move functions into utils, move sub ood methods
